Multithreading for CPU kernels
When a CPU kernel is executed, the image is now divided into tiles and each is calculated concurrently on a separate thread.
It might be good to add a threading library to do this in the future.
This does not work for odd dimension images yet. TODO

(I also had to make the random generator thread safe)